core/standards/css3/image-fit/js/ua-stylesheet/svg-video/001.html

<!doctype html>
<title>svg video default style</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var svg = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
  var e = document.createElementNS('http://www.w3.org/2000/svg', 'video');
  svg.appendChild(e);
  document.body.appendChild(svg);
  assert_equals(getComputedStyle(e, '').OObjectFit, 'contain');
  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
});
</script>
